[color] rgb_t change to uint8_t

+++ src/common-components/color/include/color.h
@@ -8,9 +8,9 @@
 extern "C"{
 #endif
 typedef struct{
-    uint16_t red;
-    uint16_t green;
-    uint16_t blue;
+    uint8_t red;
+    uint8_t green;
+    uint8_t blue;
 } rgb_t;
 
 typedef struct {
